+++ src/telegram_bot_service/ui/dialogs/error_dialog.py
@@ -1,0 +1,9 @@
+from aiogram_dialog import Dialog
+from ui.windows.error_dialog_windows.error_window import (
+    window as error_window,
+)
+
+
+dialog = Dialog(
+    error_window,
+)
